extract the dependencies deserialization to a separate module
