+++ src/dsl/schema/deserialization.rs
@@ -1,6 +1,3 @@
-use std::collections::HashMap;
-
-use balena_temen::ast::*;
 use serde::de::Error;
 use serde_yaml::Mapping;
 use serde_yaml::Value;
@@ -13,6 +10,8 @@ use crate::dsl::schema::object_types::ObjectType;
 use crate::dsl::schema::object_types::RawObjectType;
 use crate::dsl::schema::Schema;
 use crate::dsl::schema::SchemaList;
+use crate::dsl::schema::when::dependencies_for_schema_list;
+use crate::dsl::schema::when::DependencyGraph;
 
 pub fn deserialize_root<E>(schema: &Value) -> Result<DocumentRoot, CompilationError>
 where
@@ -46,102 +45,6 @@ where
         dependencies: Some(dependencies),
     })
 }
-
-/// structure representing the whole DAG of dependencies between the schemas
-#[derive(Debug, Clone)]
-pub struct DependencyGraph {
-    // schema name -> its dependencies
-    all: HashMap<String, DependencyTree>,
-}
-
-impl DependencyGraph {
-    pub fn contains(&self, schema_name: &str) -> bool {
-        self.all.contains_key(schema_name)
-    }
-
-    pub fn dependencies_for(&self, schema_name: &str) -> Vec<&str> {
-        if self.contains(schema_name) {
-            self.all[schema_name].tree.iter().map(|name| name.as_ref()).collect()
-        } else {
-            vec![]
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-struct DependencyTree {
-    tree: Vec<String>, // TODO: change into actual tree
-}
-
-impl DependencyTree {
-    fn start_with(identifiers: &Identifier) -> DependencyTree {
-        let mut result = vec![];
-        for identifier in &identifiers.values {
-            match identifier {
-                IdentifierValue::Name(name) => {
-                    result.push(name.clone());
-                }
-                _ => unimplemented!(),
-            }
-        }
-        DependencyTree { tree: result }
-    }
-    // TODO: see if we need `merge_with` as well
-}
-
-impl DependencyGraph {
-    fn empty() -> DependencyGraph {
-        DependencyGraph { all: HashMap::new() }
-    }
-
-    fn push(self, name: &str, depends_on: &Expression) -> DependencyGraph {
-        let map = match self.all.get(name) {
-            None => {
-                let mut map = self.all.clone();
-                match depends_on.value {
-                    ExpressionValue::Identifier(ref identifiers) => {
-                        map.insert(name.to_string(), DependencyTree::start_with(identifiers));
-                    }
-                    // TODO:
-                    _ => unimplemented!("walking logical expression that is not just one identifier"),
-                }
-
-                map
-            }
-            Some(_previous) => {
-                // TODO:
-                unimplemented!("merging with previously seen expression")
-            }
-        };
-
-        DependencyGraph { all: map }
-    }
-}
-
-fn dependencies_for_schema_list(
-    maybe_list: Option<&SchemaList>,
-    previous_tree: DependencyGraph,
-) -> Result<DependencyGraph, CompilationError> {
-    match maybe_list {
-        None => Ok(DependencyGraph::empty()),
-        Some(list) => {
-            let mut tree = previous_tree;
-            for schema in list.entries() {
-                if let Some(when) = &schema.schema.when {
-                    tree = tree.push(&schema.name, &when);
-                }
-
-                if let Some(children) = &schema.schema.children {
-                    for named_child in children.entries() {
-                        tree = dependencies_for_schema_list(named_child.schema.children.as_ref(), tree)?;
-                    }
-                }
-            }
-            Ok(tree)
-        }
-    }
-}
-
 pub fn deserialize_schema<E>(value: &Value) -> Result<Schema, E>
 where
     E: Error,
