+++ src/output/serialization/when.rs
@@ -5,9 +5,9 @@ use serde::ser::SerializeMap;
 use serde::Serialize;
 use serde_json::Value;
 
-use crate::dsl::schema::deserialization::DependencyGraph;
 use crate::dsl::schema::NamedSchema;
 use crate::dsl::schema::SchemaList;
+use crate::dsl::schema::when::DependencyGraph;
 
 pub fn serialize_schema_list_dependencies<O, E, S>(
     schema_list: &SchemaList,
